+++ NEWS.md
@@ -1,4 +1,5 @@
 
+* New function 'proj_crs_text()' to convert from one CRS text format to another. 
 
 * No set up on load now, all handled by libproj. Package should always function, always with version 7 or above.